+++ src/commands/parseCommands.ts
@@ -5,7 +5,6 @@ import {
   VALID_TRANSFORM_TYPES
 } from "../constants.js";
 import { convertFiles } from "../converters/converter.js";
-import { checkPath } from "../converters/utils.js";
 import { ERRORS } from "../messages/errMessages.js";
 import type {
   OptionKeysWithValues,
@@ -16,6 +15,7 @@ import type {
   UtilityActions,
   UtilityActionsCheck
 } from "../types";
+import { checkPath } from "../utils/utils.js";
 import program from "./generateCommands.js";
 import { utilityActionsCorrespondenceTable } from "./programConfiguration.js";
